+++ lib/src/configuration/better_player_video_format.dart
@@ -1,0 +1,6 @@
+enum BetterPlayerVideoFormat {
+  dash,
+  hls,
+  ss,
+  other,
+}